+++ src/Pages/DetailPage/DetailBody.tsx
@@ -1,6 +1,6 @@
 import { ChevronRight, LionProfile, LocationFrom, LocationMarker } from '../../assets/svg';
 import * as S from './style';
-import createAgo from '../MainPage/Components/SingleContent/createAgo';
+import createAgo from '../util/createAgo';
 import DetailMap from './DetailMap';
 
 interface DetailBodyProps {
