+++ src/Pages/MainPage/Components/SingleContent/SingleContentHeader.tsx
